pyproject.toml
[project]
name = "isag"
version = "0.1.0"
description = "Isag — Isolated Agent. Strict containerized sandboxes for coding agents."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
isag = "isag.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["isag"]
